Modify migrations and models to use INTEGER instead of BIGINT for sqlite auto increment
This fix is temporary.

+++ app/common/data/models.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Boolean, Column, DateTime, ForeignKey, BigInteger, LargeBinary, String
+from sqlalchemy import Boolean, Column, DateTime, ForeignKey, BigInteger, LargeBinary, String, Integer
 from sqlalchemy.orm import relationship
 
 from app.common.domain.database import Base
@@ -9,7 +9,7 @@ from app.common.domain.database import Base
 class BaseEntity(Base):
     __abstract__ = True
 
-    id = Column(BigInteger, primary_key=True, index=True)
+    id = Column(Integer, primary_key=True, index=True)
     created_on = Column(DateTime, default=datetime.utcnow, nullable=False)
     updated_on = Column(DateTime, default=None, onupdate=datetime.utcnow, nullable=True)
     is_deleted = Column(Boolean, default=False, nullable=False)
@@ -36,5 +36,5 @@ class UserToken(BaseEntity):
     token = Column(String, nullable=False, index=True)
     token_type = Column(String, nullable=False)
     expiry = Column(BigInteger, nullable=False)
-    user_id = Column(BigInteger, ForeignKey("users.id"))
+    user_id = Column(Integer, ForeignKey("users.id"))
     user = relationship("User")
